live cylc data
wip: update graph
wip: initial empty graphwaits for socket message data to update graph
wip: in progress update graph
wip: update graph
wip: access workflow data

wip: watch workflow data
wip:  node and edge data from graphql query

ready to compute correct  schema for cytoscape - requires update to graph data to schema in uiserver
wip: map node and edge properties to element data
wip: subscriptions object
wip: cleanup
wip: graph component

graph initiated from workflowid in workflows/gscan screen
wip: access new data properties
wip: route clean
wip: fix typo
debug: graph data
wip: debug edge data
debug: a little closer
debug: parent and runpercent
wip: debug

icons from state working - colour from states still need to be mapped to pie chart values
live data
live cylc data
