+++ package.json
@@ -16,6 +16,7 @@
   "dependencies": {
     "apollo-boost": "^0.4.4",
     "axios-fetch": "^1.1.0",
+    "chartist": "0.11.3",
     "cytoscape": "^3.9.0",
     "cytoscape-cola": "^2.3.0",
     "cytoscape-cose-bilkent": "^4.0.0",
@@ -27,21 +28,22 @@
     "cytoscape-popper": "^1.0.4",
     "cytoscape-undo-redo": "^1.3.1",
     "dotparser": "^0.4.0",
-    "chartist": "0.11.3",
     "graphql": "^14.4.2",
     "graphql-tag": "^2.10.1",
     "jquery": "^3.4.1",
     "jshint": "^2.10.2",
     "nprogress": "^0.2.0",
     "tippy.js": "^4.3.5",
-    "vue-cytoscape": "^0.2.10",
+    "uuid-by-string": "^2.1.5",
     "vee-validate": "^2.2.15",
     "vue": "^2.6.10",
     "vue-apollo": "^3.0.0-beta.30",
     "vue-chartist": "^2.2.1",
+    "vue-cytoscape": "^0.2.10",
     "vue-flag-icon": "^1.0.6",
     "vue-i18n": "^8.14.0",
     "vue-meta": "^2.2.1",
+    "vue-native-websocket": "^2.0.13",
     "vue-router": "^3.1.2",
     "vue-spinner": "^1.0.3",
     "vuetify": "1.5.16",
