--- public/simple-cytoscape-dot.7.js
@@ -1,1119 +1,0 @@
-{
-  "version": 101,
-  "nodes": [
-    {
-      "data": {
-        "id": "338c432b-9f96-420c-908f-84d06539cb1d",
-        "parent": "",
-        "label": "long_start_12.20150219T1200Z",
-        "suid": 442,
-        "shape": "ellipse",
-        "state": "#31ff53",
-        "expired": 0,
-        "queued": 0,
-        "retrying": 0,
-        "waiting": 0,
-        "running": 0,
-        "submitted": 0,
-        "succeeded": 100,
-        "failed": 0,
-        "subfailed": 0,
-        "todo": 0,
-        "icon": "/img/outline-radio_button_unchecked-24px.svg"
-      },
-      "position":{
-
-      },
-      "group":"nodes",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "946be865-95c9-4cf1-89f5-a06400f739b9",
-        "parent": "",
-        "label": "long_res_create.20150219T1200Z",
-        "suid": 441,
-        "shape": "ellipse",
-        "state": "#9ef9ff",
-        "expired": 0,
-        "queued": 0,
-        "retrying": 0,
-        "waiting": 0,
-        "running": 0,
-        "submitted": 100,
-        "succeeded": 0,
-        "failed": 0,
-        "subfailed": 0,
-        "todo": 0,
-        "icon": "/img/outline-adjust-24px.svg"
-      },
-      "position":{
-
-      },
-      "group":"nodes",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "a411f6ea-eb02-45e9-94bf-65afc49f19a3",
-        "parent": "",
-        "label": "long_start_06.20150219T0600Z",
-        "suid": 440,
-        "shape": "ellipse",
-        "state": "grey",
-        "expired": 0,
-        "queued": 0,
-        "retrying": 0,
-        "waiting": 0,
-        "running": 90,
-        "submitted": 0,
-        "succeeded": 0,
-        "failed": 0,
-        "subfailed": 0,
-        "todo": 10,
-        "icon": "/img/outline-adjust-24px.svg"
-      },
-      "position":{
-
-      },
-      "group":"nodes",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "8c403369-09b7-432c-97eb-c29606c11c6f",
-        "parent": "",
-        "label": "long_res_create.20150219T0600Z",
-        "suid": 439,
-        "shape": "ellipse",
-        "state": "#31ff53",
-        "expired": 0,
-        "queued": 0,
-        "retrying": 0,
-        "waiting": 0,
-        "running": 0,
-        "submitted": 0,
-        "succeeded": 100,
-        "failed": 0,
-        "subfailed": 0,
-        "todo": 0,
-        "icon": "/img/outline-radio_button_unchecked-24px.svg"
-      },
-      "position":{
-
-      },
-      "group":"nodes",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "f62109bf-d37e-45c0-9371-5a7859c6404c",
-        "parent": "",
-        "label": "long_start_00.20150219T0000Z",
-        "suid": 438,
-        "shape": "ellipse",
-        "state": "#ff3a2b",
-        "expired": 0,
-        "queued": 0,
-        "retrying": 0,
-        "waiting": 0,
-        "running": 0,
-        "submitted": 0,
-        "succeeded": 0,
-        "failed": 100,
-        "subfailed": 0,
-        "todo": 0,
-        "icon": "/img/outline-cancel-24px.svg"
-      },
-      "position":{
-
-      },
-      "group":"nodes",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "868d195d-95ea-482f-815d-410c28b3252e",
-        "parent": "",
-        "label": "long_res_create.20150219T0000Z",
-        "suid": 437,
-        "shape": "ellipse",
-        "state": "#d453ff",
-        "expired": 0,
-        "queued": 0,
-        "retrying": 0,
-        "waiting": 0,
-        "running": 0,
-        "submitted": 0,
-        "succeeded": 0,
-        "failed": 0,
-        "subfailed": 100,
-        "todo": 0,
-        "icon": "/img/outline-filter_tilt_shift-24px.svg"
-      },
-      "position":{
-
-      },
-      "group":"nodes",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "435fe1b0-25c8-4c05-8f99-cee1c4299686",
-        "parent": "",
-        "label": "long_forecast_create_zap_little.20150219T1200Z",
-        "suid": 436,
-        "shape": "ellipse",
-        "state": "#31ff53",
-        "expired": 0,
-        "queued": 0,
-        "retrying": 0,
-        "waiting": 0,
-        "running": 0,
-        "submitted": 0,
-        "succeeded": 100,
-        "failed": 0,
-        "subfailed": 0,
-        "todo": 0,
-        "icon": "/img/outline-radio_button_unchecked-24px.svg"
-      },
-      "position":{
-
-      },
-      "group":"nodes",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "10cb9b20-f73e-48ad-bbe7-2001b64a63ed",
-        "parent": "",
-        "label": "long_forecast_create_zap_larger.20150219T1200Z",
-        "suid": 435,
-        "shape": "ellipse",
-        "state": "#fff138",
-        "expired": 0,
-        "queued": 100,
-        "retrying": 0,
-        "waiting": 0,
-        "running": 0,
-        "submitted": 0,
-        "succeeded": 0,
-        "failed": 0,
-        "subfailed": 0,
-        "todo": 0,
-        "icon": "/img/baseline-donut_large-24px.svg"
-      },
-      "position":{
-
-      },
-      "group":"nodes",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "82d15bc9-897b-464a-b367-95167c102f16",
-        "parent": "",
-        "label": "recover3.20150219T1200Z",
-        "suid": 434,
-        "shape": "ellipse",
-        "state": "#ccc",
-        "expired": 0,
-        "queued": 0,
-        "retrying": 0,
-        "waiting": 0,
-        "running": 20,
-        "submitted": 0,
-        "succeeded": 0,
-        "failed": 0,
-        "subfailed": 0,
-        "todo": 80,
-        "icon": "/img/outline-adjust-24px.svg"
-      },
-      "position":{
-
-      },
-      "group":"nodes",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "7e9b1828-07f2-4171-b24e-3440500c4941",
-        "parent": "",
-        "label": "recover3.20150219T0600Z",
-        "suid": 433,
-        "shape": "ellipse",
-        "state": "#9ef9ff",
-        "expired": 0,
-        "queued": 0,
-        "retrying": 0,
-        "waiting": 0,
-        "running": 0,
-        "submitted": 100,
-        "succeeded": 0,
-        "failed": 0,
-        "subfailed": 0,
-        "todo": 0,
-        "icon": "/img/outline-adjust-24px.svg"
-      },
-      "position":{
-
-      },
-      "group":"nodes",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "80bfca4b-506d-4d5f-82aa-c999509f0dd0",
-        "parent": "",
-        "label": "recover3.20150219T0000Z",
-        "suid": 432,
-        "shape": "ellipse",
-        "state": "#31ff53",
-        "expired": 0,
-        "queued": 0,
-        "retrying": 0,
-        "waiting": 0,
-        "running": 0,
-        "submitted": 0,
-        "succeeded": 100,
-        "failed": 0,
-        "subfailed": 0,
-        "todo": 0,
-        "icon": "/img/outline-radio_button_unchecked-24px.svg"
-      },
-      "position":{
-
-      },
-      "group":"nodes",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "0696f731-4522-4f01-9bf1-b5d035a117a2",
-        "parent": "",
-        "label": "long_assimilation_create_zap_little.20150219T1200Z",
-        "suid": 431,
-        "shape": "ellipse",
-        "state": "grey",
-        "expired": 0,
-        "queued": 0,
-        "retrying": 0,
-        "waiting": 0,
-        "running": 0,
-        "submitted": 0,
-        "succeeded": 0,
-        "failed": 0,
-        "subfailed": 0,
-        "todo": 0,
-        "icon": "/img/baseline-donut_large-24px.svg"
-      },
-      "position":{
-
-      },
-      "group":"nodes",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "c51658c1-4466-4093-8a24-cef092c7dc7f",
-        "parent": "0696f731-4522-4f01-9bf1-b5d035a117a2",
-        "label": "long_sync_forecast_tJ.20150219T0600Z",
-        "suid": 1000,
-        "shape": "ellipse",
-        "state": "#fff138",
-        "expired": 0,
-        "queued": 100,
-        "retrying": 0,
-        "waiting": 0,
-        "running": 0,
-        "submitted": 0,
-        "succeeded": 0,
-        "failed": 0,
-        "subfailed": 0,
-        "todo": 0,
-        "icon": "/img/baseline-donut_large-24px.svg"
-      },
-      "position":{
-
-      },
-      "group":"nodes",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "464a09e4-8b01-4c46-9718-4b14e979cde2",
-        "parent": "0696f731-4522-4f01-9bf1-b5d035a117a2",
-        "label": "long_sync_forecast.20150219T0600Z",
-        "suid": 1001,
-        "shape": "ellipse",
-        "state": "#fff",
-        "expired": 0,
-        "queued": 0,
-        "retrying": 0,
-        "waiting": 100,
-        "running": 0,
-        "submitted": 0,
-        "succeeded": 0,
-        "failed": 0,
-        "subfailed": 100,
-        "todo": 0,
-        "icon": "/img/outline-radio_button_unchecked-24px.svg"
-      },
-      "position":{
-
-      },
-      "group":"nodes",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "d931ddf4-fe6e-4f23-9562-eb078777a3e0",
-        "parent": "0696f731-4522-4f01-9bf1-b5d035a117a2",
-        "label": "long_send_stingray_stingray.20150219T0600Z",
-        "suid": 1002,
-        "shape": "ellipse",
-        "state": "#fff138",
-        "expired": 0,
-        "queued": 100,
-        "retrying": 0,
-        "waiting": 0,
-        "running": 0,
-        "submitted": 0,
-        "succeeded": 0,
-        "failed": 0,
-        "subfailed": 0,
-        "todo": 0,
-        "icon": "/img/baseline-donut_large-24px.svg"
-      },
-      "position":{
-
-      },
-      "group":"nodes",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "af65e13d-9876-41fc-85ef-02f62af80b24",
-        "parent": "0696f731-4522-4f01-9bf1-b5d035a117a2",
-        "label": "long_res_delete.20150219T0600Z",
-        "suid": 1003,
-        "shape": "ellipse",
-        "state": "#9ef9ff",
-        "expired": 0,
-        "queued": 0,
-        "retrying": 0,
-        "waiting": 0,
-        "running": 0,
-        "submitted": 100,
-        "succeeded": 0,
-        "failed": 0,
-        "subfailed": 0,
-        "todo": 0,
-        "icon": "/img/outline-adjust-24px.svg"
-      },
-      "position":{
-
-      },
-      "group":"nodes",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "cc44a0ed-c24f-42fa-a6a5-ed28c1825220",
-        "parent": "0696f731-4522-4f01-9bf1-b5d035a117a2",
-        "label": "long_happyland_ingest.20150219T0600Z",
-        "suid": 1004,
-        "shape": "ellipse",
-        "state": "#ff3a2b",
-        "expired": 0,
-        "queued": 0,
-        "retrying": 100,
-        "waiting": 0,
-        "running": 0,
-        "submitted": 0,
-        "succeeded": 0,
-        "failed": 10,
-        "subfailed": 0,
-        "todo": 0,
-        "icon": "/img/outline-refresh-24px.svg"
-      },
-      "position":{
-
-      },
-      "group":"nodes",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "13ed33c5-5003-4d41-95cc-770ef6fdedb0",
-        "parent": "c51658c1-4466-4093-8a24-cef092c7dc7f",
-        "label": "ENSEMBLE_SUBJOB.20150219T0000Z (interaction) ensemble_subjob_hook.20150219T0000Z",
-        "suid": 1005,
-        "shape": "ellipse",
-        "state": "#fff",
-        "expired": 0,
-        "queued": 0,
-        "retrying": 0,
-        "waiting": 100,
-        "running": 0,
-        "submitted": 0,
-        "succeeded": 0,
-        "failed": 0,
-        "subfailed": 0,
-        "todo": 0,
-        "icon": "/img/outline-radio_button_unchecked-24px.svg"
-      },
-      "position":{
-
-      },
-      "group":"nodes",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "fbc7ebd3-c937-4ff3-9c48-94d18d82bdeb",
-        "parent": "c51658c1-4466-4093-8a24-cef092c7dc7f",
-        "label": "brief_sync_happy.20150219T0600Z",
-        "suid": 1006,
-        "shape": "ellipse",
-        "state": "#9ef9ff",
-        "expired": 0,
-        "queued": 0,
-        "retrying": 0,
-        "waiting": 0,
-        "running": 0,
-        "submitted": 100,
-        "succeeded": 0,
-        "failed": 0,
-        "subfailed": 0,
-        "todo": 0,
-        "icon": "/img/outline-adjust-24px.svg"
-      },
-      "position":{
-
-      },
-      "group":"nodes",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "d03ba647-cb9b-46c5-b040-51c0c0b42969",
-        "parent": "c51658c1-4466-4093-8a24-cef092c7dc7f",
-        "label": "brief_sync_forecast.20150219T0600Z",
-        "suid": 1007,
-        "shape": "ellipse",
-        "state": "#ff3a2b",
-        "expired": 0,
-        "queued": 0,
-        "retrying": 0,
-        "waiting": 0,
-        "running": 0,
-        "submitted": 0,
-        "succeeded": 0,
-        "failed": 100,
-        "subfailed": 0,
-        "todo": 0,
-        "icon": "/img/outline-cancel-24px.svg"
-      },
-      "position":{
-
-      },
-      "group":"nodes",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "f9257ea7-9da7-44f7-84e8-11911bac1e24",
-        "parent": "",
-        "label": "archive_logs.20150219T0000Z",
-        "suid": 1008,
-        "shape": "ellipse",
-        "state": "grey",
-        "expired": 0,
-        "queued": 0,
-        "retrying": 0,
-        "waiting": 0,
-        "running": 0,
-        "submitted": 0,
-        "succeeded": 0,
-        "failed": 0,
-        "subfailed": 0,
-        "todo": 0,
-        "icon": ""
-      },
-      "position":{
-
-      },
-      "group":"nodes",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "b5766bff-331d-4e5e-a71c-cbd3a6005324",
-        "parent": "",
-        "label": "ENSEMBLE_SUBJOB.20150219T0600Z",
-        "suid": 1009,
-        "shape": "ellipse",
-        "state": "grey",
-        "expired": 0,
-        "queued": 0,
-        "retrying": 0,
-        "waiting": 0,
-        "running": 0,
-        "submitted": 0,
-        "succeeded": 0,
-        "failed": 0,
-        "subfailed": 0,
-        "todo": 0,
-        "icon": ""
-      },
-      "position":{
-
-      },
-      "group":"nodes",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "f7982efe-6aab-47e0-a343-40587a17e772",
-        "parent": "",
-        "label": "ENSEMBLE_VERIFICATIONMODEL.20150219T0600Z",
-        "suid": 1010,
-        "shape": "ellipse",
-        "state": "grey",
-        "expired": 0,
-        "queued": 0,
-        "retrying": 0,
-        "waiting": 0,
-        "running": 0,
-        "submitted": 0,
-        "succeeded": 0,
-        "failed": 0,
-        "subfailed": 0,
-        "todo": 0,
-        "icon": ""
-      },
-      "position":{
-
-      },
-      "group":"nodes",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    }
-  ],
-  "edges": [
-    {
-      "data": {
-        "id": "1095",
-        "source": "af65e13d-9876-41fc-85ef-02f62af80b24",
-        "target": "f7982efe-6aab-47e0-a343-40587a17e772",
-        "label": "ensemble_thunderbirds_are_go_forecast_002.20150219T1200Z (interaction) ensemble_sync_happy.20150219T1200Z",
-        "suid": 1095
-      },
-      "position":{
-
-      },
-      "group":"edges",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "1094",
-        "source": "af65e13d-9876-41fc-85ef-02f62af80b24",
-        "target": "b5766bff-331d-4e5e-a71c-cbd3a6005324",
-        "label": "ensemble_thunderbirds_are_go_forecast_003.20150219T1200Z (interaction) ensemble_sync_happy.20150219T1200Z",
-        "suid": 1094
-      },
-      "position":{
-
-      },
-      "group":"edges",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-     {
-      "data": {
-        "id": "1093",
-        "source": "af65e13d-9876-41fc-85ef-02f62af80b24",
-        "target": "f9257ea7-9da7-44f7-84e8-11911bac1e24",
-        "label": "brief_forecast.20150219T0600Z (interaction) brief_verificationmodel_database.20150219T0600Z",
-        "suid": 1093
-      },
-      "position":{
-
-      },
-      "group":"edges",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "1092",
-        "source": "c51658c1-4466-4093-8a24-cef092c7dc7f",
-        "target": "d03ba647-cb9b-46c5-b040-51c0c0b42969",
-        "label": "BRIEF_SURFACE_LANDSIM.20150219T0600Z (interaction) brief_wagtail_filter.20150219T0600Z",
-        "suid": 1092
-      },
-      "position":{
-
-      },
-      "group":"edges",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "1091",
-        "source": "c51658c1-4466-4093-8a24-cef092c7dc7f",
-        "target": "fbc7ebd3-c937-4ff3-9c48-94d18d82bdeb",
-        "label": "long_start.20150219T0000Z (interaction) long_start.20150219T0600Z",
-        "suid": 1091
-      },
-      "position":{
-
-      },
-      "group":"edges",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "1090",
-        "source": "c51658c1-4466-4093-8a24-cef092c7dc7f",
-        "target": "13ed33c5-5003-4d41-95cc-770ef6fdedb0",
-        "label": "brief_observations_background_atmos.20150219T0000Z (interaction) brief_observations_merge_background.20150219T0000Z",
-        "suid": 1090
-      },
-      "position":{
-
-      },
-      "group":"edges",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "1089",
-        "source": "82d15bc9-897b-464a-b367-95167c102f16",
-        "target": "cc44a0ed-c24f-42fa-a6a5-ed28c1825220",
-        "label": "ENSEMBLE_SUBJOB.20150219T0000Z (interaction) ensemble_thunderbirds_are_go_forecast_000.20150219T0000Z",
-        "suid": 1089
-      },
-      "position":{
-
-      },
-      "group":"edges",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "1088",
-        "source": "82d15bc9-897b-464a-b367-95167c102f16",
-        "target": "af65e13d-9876-41fc-85ef-02f62af80b24",
-        "label": "long_observations_background_atmos.20150219T1200Z (interaction) LONG_OBS_PROCESS_BACKGROUND.20150219T1200Z",
-        "suid": 1088
-      },
-      "position":{
-
-      },
-      "group":"edges",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "1087",
-        "source": "82d15bc9-897b-464a-b367-95167c102f16",
-        "target": "d931ddf4-fe6e-4f23-9562-eb078777a3e0",
-        "label": "BRIEF_SURFACE_LANDSIM.20150219T1200Z (interaction) brief_wagtail_filter.20150219T1200Z",
-        "suid": 1087
-      },
-      "position":{
-
-      },
-      "group":"edges",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "1086",
-        "source": "82d15bc9-897b-464a-b367-95167c102f16",
-        "target": "464a09e4-8b01-4c46-9718-4b14e979cde2",
-        "label": "brief_observations_background_atmos.20150219T1200Z (interaction) BRIEF_OBSERVATIONS_PROCESS_BACKGROUND.20150219T1200Z",
-        "suid": 1086
-      },
-      "position":{
-
-      },
-      "group":"edges",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "1085",
-        "source": "82d15bc9-897b-464a-b367-95167c102f16",
-        "target": "c51658c1-4466-4093-8a24-cef092c7dc7f",
-        "label": "brief_observations_background_atmos.20150219T1200Z (interaction) brief_observations_merge_background.20150219T1200Z",
-        "suid": 1085
-      },
-      "position":{
-
-      },
-      "group":"edges",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "1084",
-        "source": "338c432b-9f96-420c-908f-84d06539cb1d",
-        "target": "10cb9b20-f73e-48ad-bbe7-2001b64a63ed",
-        "label": "long_start_12.20150219T1200Z (interaction) long_start.20150219T1200Z",
-        "suid": 1084
-      },
-      "position":{
-
-      },
-      "group":"edges",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "1083",
-        "source": "946be865-95c9-4cf1-89f5-a06400f739b9",
-        "target": "338c432b-9f96-420c-908f-84d06539cb1d",
-        "label": "long_res_create.20150219T1200Z (interaction) long_start_12.20150219T1200Z",
-        "suid": 1083
-      },
-      "position":{
-
-      },
-      "group":"edges",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "1082",
-        "source": "a411f6ea-eb02-45e9-94bf-65afc49f19a3",
-        "target": "338c432b-9f96-420c-908f-84d06539cb1d",
-        "label": "long_start_06.20150219T0600Z (interaction) long_start.20150219T0600Z",
-        "suid": 1082
-      },
-      "position":{
-
-      },
-      "group":"edges",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "1081",
-        "source": "8c403369-09b7-432c-97eb-c29606c11c6f",
-        "target": "a411f6ea-eb02-45e9-94bf-65afc49f19a3",
-        "label": "long_res_create.20150219T0600Z (interaction) long_start_06.20150219T0600Z",
-        "suid": 1081
-      },
-      "position":{
-
-      },
-      "group":"edges",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "1080",
-        "source": "f62109bf-d37e-45c0-9371-5a7859c6404c",
-        "target": "82d15bc9-897b-464a-b367-95167c102f16",
-        "label": "long_start_00.20150219T0000Z (interaction) long_start.20150219T0000Z",
-        "suid": 1080
-      },
-      "position":{
-
-      },
-      "group":"edges",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "1079",
-        "source": "868d195d-95ea-482f-815d-410c28b3252e",
-        "target": "f62109bf-d37e-45c0-9371-5a7859c6404c",
-        "label": "long_res_create.20150219T0000Z (interaction) long_start_00.20150219T0000Z",
-        "suid": 1079
-      },
-      "position":{
-
-      },
-      "group":"edges",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "1078",
-        "source": "435fe1b0-25c8-4c05-8f99-cee1c4299686",
-        "target": "82d15bc9-897b-464a-b367-95167c102f16",
-        "label": "long_forecast_create_zap_little.20150219T1200Z (interaction) long_assimilation_create_zap_little.20150219T1200Z",
-        "suid": 1078
-      },
-      "position":{
-
-      },
-      "group":"edges",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "1076",
-        "source": "10cb9b20-f73e-48ad-bbe7-2001b64a63ed",
-        "target": "82d15bc9-897b-464a-b367-95167c102f16",
-        "label": "long_forecast_create_zap_larger.20150219T1200Z (interaction) long_assimilation_create_zap_larger.20150219T1200Z",
-        "suid": 1076
-      },
-      "position":{
-
-      },
-      "group":"edges",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "1099",
-        "source": "7e9b1828-07f2-4171-b24e-3440500c4941",
-        "target": "8c403369-09b7-432c-97eb-c29606c11c6f",
-        "label": "recover3.20150219T0600Z (interaction) brief_forecast.20150219T0600Z",
-        "suid": 1099
-      },
-      "position":{
-
-      },
-      "group":"edges",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    },
-    {
-      "data": {
-        "id": "1098",
-        "source": "80bfca4b-506d-4d5f-82aa-c999509f0dd0",
-        "target": "8c403369-09b7-432c-97eb-c29606c11c6f",
-        "label": "recover3.20150219T0000Z (interaction) brief_forecast.20150219T0000Z",
-        "suid": 1098
-      },
-      "position":{
-
-      },
-      "group":"edges",
-      "removed":false,
-      "selected":false,
-      "selectable":true,
-      "locked":false,
-      "grabbable":true,
-      "classes":""
-    }
-  ]
-}
