+++ src/services/graph.service.js
@@ -1,0 +1,14 @@
+/* jshint esversion: 6, asi: true */
+// import Vue from 'vue'
+// import store from '@/store/'
+// import VueNativeSock from 'vue-native-websocket'
+
+// const SOCKET_SERVER = 'wss://echo.websocket.org'
+
+// Vue.use(VueNativeSock, SOCKET_SERVER, {
+//   store: store.graph,
+//   format: 'json',
+//   reconnection: true, // (Boolean) whether to reconnect automatically (false)
+//   reconnectionAttempts: 5, // (Number) number of reconnection attempts before giving up (Infinity),
+//   reconnectionDelay: 3000 // (Number) how long to initially wait before attempting a new (1000)
+// })
