+++ src/styles/cytoscape/cytoscape-custom.css
@@ -12,20 +12,6 @@
   z-index: 2;
 }
 
-#cytoscape { 
-  width: 100%;
-  height: 800px;
-  background-color: #222;
-  z-index: 100;
-}
-
-#cytoscape { 
-  width: 100%;
-  height: 800px;
-  background-color: #222;
-  z-index: 100;
-}
-
 #app {
   font-family: 'Avenir', Helvetica, Arial, sans-serif;
   -webkit-font-smoothing: antialiased;
